fix: auto get token, when token expired
