+++ src/pageComponents/WithdrawContent/index.tsx
@@ -77,7 +77,8 @@ import RemainingQuato from './RemainingQuato';
 // import { getRecordStatus } from 'utils/api/records';
 import { setIsShowRedDot } from 'store/reducers/common/slice';
 import { useWalletContext } from 'provider/walletProvider';
-import { isHtmlError } from 'utils/api/error';
+import { isAuthTokenError, isHtmlError } from 'utils/api/error';
+import myEvents from 'utils/myEvent';
 
 enum ValidateStatus {
   Error = 'error',
@@ -337,7 +338,7 @@ export default function WithdrawContent() {
               errorMessage: '',
             },
           });
-          if (error.name !== CommonErrorNameType.CANCEL) {
+          if (error.name !== CommonErrorNameType.CANCEL && !isAuthTokenError(error)) {
             singleMessage.error(handleErrorMessage(error));
           }
           setNetworkList([]);
@@ -447,7 +448,8 @@ export default function WithdrawContent() {
         setWithdrawInfo({ ...InitialWithdrawInfo, transactionUnit: symbol });
         if (
           error.name !== CommonErrorNameType.CANCEL &&
-          !isHtmlError(error?.code, handleErrorMessage(error))
+          !isHtmlError(error?.code, handleErrorMessage(error)) &&
+          !isAuthTokenError(error)
         ) {
           singleMessage.error(handleErrorMessage(error));
           setIsTransactionFeeLoading(false);
@@ -837,7 +839,7 @@ export default function WithdrawContent() {
     getWithdrawData();
   }, [form, getWithdrawData]);
 
-  useEffectOnce(() => {
+  const init = useCallback(() => {
     form.setFieldValue(FormKeys.ADDRESS, withdraw.address || '');
 
     if (
@@ -854,6 +856,17 @@ export default function WithdrawContent() {
     } else {
       handleChainChanged(currentChainItemRef.current);
     }
+  }, [
+    form,
+    getWithdrawData,
+    handleChainChanged,
+    withdraw.address,
+    withdraw.currentNetwork,
+    withdraw.networkList,
+  ]);
+
+  useEffectOnce(() => {
+    init();
 
     // interval fetch balance
     getMaxBalanceInterval();
@@ -861,6 +874,16 @@ export default function WithdrawContent() {
       if (getMaxBalanceTimerRef.current) clearInterval(getMaxBalanceTimerRef.current);
     };
   });
+
+  useEffect(() => {
+    const { remove } = myEvents.AuthTokenSuccess.addListener(() => {
+      console.log('login success');
+      init();
+    });
+    return () => {
+      remove();
+    };
+  }, [init]);
 
   const renderTransactionFeeValue = () => {
     if (!withdrawInfo.transactionFee || !withdrawInfo.aelfTransactionFee) {
