+++ src/provider/initProviderV1.ts
@@ -13,7 +13,7 @@ import { useEffectOnce } from 'react-use';
 import { useThrottleCallback } from 'hooks';
 import { usePortkeyProvider } from 'hooks/usePortkeyProvider';
 import myEvents from 'utils/myEvent';
-import { resetJWT } from 'api/utils';
+import { resetLocalJWT } from 'api/utils';
 import singleMessage from 'components/SingleMessage';
 import 'utils/firebase';
 import { LoginExpiredTip, NetworkNotMatchTipPrefix } from 'constants/wallet';
@@ -92,7 +92,7 @@ export default function InitProvider() {
   useEffectOnce(() => {
     const listener = myEvents.DeniedRequest.addListener(() => {
       singleMessage.error(LoginExpiredTip);
-      resetJWT();
+      resetLocalJWT();
       store.dispatch(setV1DisconnectedAction());
       portkeyWallet.clearData();
       resetStore();
