+++ src/utils/aelfUtils.tsx
@@ -554,3 +554,9 @@ export const getBalance = async ({
   });
   return res?.balance;
 };
+
+export const pubKeyToAddress = (pubKey: string) => {
+  const onceSHAResult = Buffer.from(AElf.utils.sha256(Buffer.from(pubKey, 'hex')), 'hex');
+  const hash = AElf.utils.sha256(onceSHAResult).slice(0, 64);
+  return AElf.utils.encodeAddressRep(hash);
+};
