+++ src/utils/etransferInstance.ts
@@ -1,0 +1,13 @@
+class ETransferInstance {
+  public obtainingToken: boolean;
+
+  constructor() {
+    this.obtainingToken = false;
+  }
+
+  setObtainingToken(status: boolean) {
+    this.obtainingToken = status;
+  }
+}
+
+export const eTransferInstance = new ETransferInstance();
